+++ services/alb/model_load_balancer.go
@@ -100,6 +100,26 @@ func setLoadBalancerGetListenersAttributeType(arg *LoadBalancerGetListenersAttri
 }
 
 /*
+	types and functions for loadBalancerSecurityGroup
+*/
+
+// isModel
+type LoadBalancerGetLoadBalancerSecurityGroupAttributeType = *CreateLoadBalancerPayloadLoadBalancerSecurityGroup
+type LoadBalancerGetLoadBalancerSecurityGroupArgType = CreateLoadBalancerPayloadLoadBalancerSecurityGroup
+type LoadBalancerGetLoadBalancerSecurityGroupRetType = CreateLoadBalancerPayloadLoadBalancerSecurityGroup
+
+func getLoadBalancerGetLoadBalancerSecurityGroupAttributeTypeOk(arg LoadBalancerGetLoadBalancerSecurityGroupAttributeType) (ret LoadBalancerGetLoadBalancerSecurityGroupRetType, ok bool) {
+	if arg == nil {
+		return ret, false
+	}
+	return *arg, true
+}
+
+func setLoadBalancerGetLoadBalancerSecurityGroupAttributeType(arg *LoadBalancerGetLoadBalancerSecurityGroupAttributeType, val LoadBalancerGetLoadBalancerSecurityGroupRetType) {
+	*arg = &val
+}
+
+/*
 	types and functions for name
 */
 
@@ -423,7 +443,8 @@ type LoadBalancer struct {
 	// External application load balancer IP address where this application load balancer is exposed. Not changeable after creation.
 	ExternalAddress LoadBalancerGetExternalAddressAttributeType `json:"externalAddress,omitempty"`
 	// There is a maximum listener count of 20.
-	Listeners LoadBalancerGetListenersAttributeType `json:"listeners,omitempty"`
+	Listeners                 LoadBalancerGetListenersAttributeType                 `json:"listeners,omitempty"`
+	LoadBalancerSecurityGroup LoadBalancerGetLoadBalancerSecurityGroupAttributeType `json:"loadBalancerSecurityGroup,omitempty"`
 	// Application Load Balancer name. Not changeable after creation.
 	Name LoadBalancerGetNameAttributeType `json:"name,omitempty"`
 	// List of networks that listeners and targets reside in. Currently limited to one. Not changeable after creation.
@@ -550,6 +571,29 @@ func (o *LoadBalancer) HasListeners() bool {
 // SetListeners gets a reference to the given []Listener and assigns it to the Listeners field.
 func (o *LoadBalancer) SetListeners(v LoadBalancerGetListenersRetType) {
 	setLoadBalancerGetListenersAttributeType(&o.Listeners, v)
+}
+
+// GetLoadBalancerSecurityGroup returns the LoadBalancerSecurityGroup field value if set, zero value otherwise.
+func (o *LoadBalancer) GetLoadBalancerSecurityGroup() (res LoadBalancerGetLoadBalancerSecurityGroupRetType) {
+	res, _ = o.GetLoadBalancerSecurityGroupOk()
+	return
+}
+
+// GetLoadBalancerSecurityGroupOk returns a tuple with the LoadBalancerSecurityGroup field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *LoadBalancer) GetLoadBalancerSecurityGroupOk() (ret LoadBalancerGetLoadBalancerSecurityGroupRetType, ok bool) {
+	return getLoadBalancerGetLoadBalancerSecurityGroupAttributeTypeOk(o.LoadBalancerSecurityGroup)
+}
+
+// HasLoadBalancerSecurityGroup returns a boolean if a field has been set.
+func (o *LoadBalancer) HasLoadBalancerSecurityGroup() bool {
+	_, ok := o.GetLoadBalancerSecurityGroupOk()
+	return ok
+}
+
+// SetLoadBalancerSecurityGroup gets a reference to the given CreateLoadBalancerPayloadLoadBalancerSecurityGroup and assigns it to the LoadBalancerSecurityGroup field.
+func (o *LoadBalancer) SetLoadBalancerSecurityGroup(v LoadBalancerGetLoadBalancerSecurityGroupRetType) {
+	setLoadBalancerGetLoadBalancerSecurityGroupAttributeType(&o.LoadBalancerSecurityGroup, v)
 }
 
 // GetName returns the Name field value if set, zero value otherwise.
@@ -795,6 +839,9 @@ func (o LoadBalancer) ToMap() (map[string]interface{}, error) {
 	}
 	if val, ok := getLoadBalancerGetListenersAttributeTypeOk(o.Listeners); ok {
 		toSerialize["Listeners"] = val
+	}
+	if val, ok := getLoadBalancerGetLoadBalancerSecurityGroupAttributeTypeOk(o.LoadBalancerSecurityGroup); ok {
+		toSerialize["LoadBalancerSecurityGroup"] = val
 	}
 	if val, ok := getLoadBalancerGetNameAttributeTypeOk(o.Name); ok {
 		toSerialize["Name"] = val